+++ src/app/model/Rook.ts
@@ -3,7 +3,7 @@ import { Piece } from './Piece';
 import { PieceMovement } from './PieceMovement';
 import { Side, Square } from './Types';
 
-export class Rock extends Piece implements PieceMovement {
+export class Rook extends Piece implements PieceMovement {
   readonly boardSize = 8;
 
   constructor(id: string, side: Side) {
@@ -28,12 +28,11 @@ export class Rock extends Piece implements PieceMovement {
       return movesToGo;
     } else {
       for (let i = position.row; i < this.boardSize; i++) {
-        if (board.squares[i][position.column - 1].piece == null) {
-          movesToGo.push(board.squares[i][position.column - 1]);
+        if (board.squares[i][position.column.indexOf(position.column)].piece == null) {
+          movesToGo.push(board.squares[i][position.column.indexOf(position.column)]);
         } else {
-          if (this.checkIfOponent(i, position.column - 1, board, this.side)) {
-            // if (board.squares[i][position.column - 1].piece?.side !== this.side) {
-            movesToGo.push(board.squares[i][position.column - 1]);
+          if (this.checkIfOponent(i, position.column.indexOf(position.column), board, this.side)) {
+            movesToGo.push(board.squares[i][position.column.indexOf(position.column)]);
             break;
           } else break;
         }
@@ -50,12 +49,11 @@ export class Rock extends Piece implements PieceMovement {
       return movesToGo;
     } else {
       for (let i = position.row - 2; i >= 0; i--) {
-        if (board.squares[i][position.column - 1].piece == null) {
-          movesToGo.push(board.squares[i][position.column - 1]);
+        if (board.squares[i][position.column.indexOf(position.column)].piece == null) {
+          movesToGo.push(board.squares[i][position.column.indexOf(position.column)]);
         } else {
-          if (this.checkIfOponent(i, position.column - 1, board, this.side)) {
-            // if (board.squares[i][position.column - 1].piece?.side !== this.side) {
-            movesToGo.push(board.squares[i][position.column - 1]);
+          if (this.checkIfOponent(i, position.column.indexOf(position.column), board, this.side)) {
+            movesToGo.push(board.squares[i][position.column.indexOf(position.column)]);
             break;
           } else break;
         }
@@ -68,15 +66,14 @@ export class Rock extends Piece implements PieceMovement {
   private goLeft(position: Square, board: Board): Square[] {
     const movesToGo: Square[] = [];
 
-    if (position.column === 1) {
+    if (position.column.indexOf(position.column) === 1) {
       return movesToGo;
     } else {
-      for (let i = position.column - 2; i >= 0; i--) {
+      for (let i = position.column.indexOf(position.column) - 1; i >= 0; i--) {
         if (board.squares[position.row - 1][i].piece == null) {
           movesToGo.push(board.squares[position.row - 1][i]);
         } else {
           if (this.checkIfOponent(position.row - 1, i, board, this.side)) {
-            // if (board.squares[position.row - 1][i].piece?.side !== this.side) {
             movesToGo.push(board.squares[position.row - 1][i]);
             break;
           } else break;
@@ -90,15 +87,14 @@ export class Rock extends Piece implements PieceMovement {
   private goRight(position: Square, board: Board): Square[] {
     const movesToGo: Square[] = [];
 
-    if (position.column === 8) {
+    if (position.column.indexOf(position.column) === 8) {
       return movesToGo;
     } else {
-      for (let i = position.column; i < this.boardSize; i++) {
+      for (let i = position.column.indexOf(position.column); i < this.boardSize; i++) {
         if (board.squares[position.row - 1][i].piece == null) {
           movesToGo.push(board.squares[position.row - 1][i]);
         } else {
           if (this.checkIfOponent(position.row - 1, i, board, this.side)) {
-            // if (board.squares[position.row - 1][i].piece?.side !== this.side) {
             movesToGo.push(board.squares[position.row - 1][i]);
             break;
           } else break;
